LayoutTests/fast/dom/html-link-element-activation-behavior-on-mouse-click-expected.txt

This tests that html link element has no activation behavior on mouse click

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.data is "test:ok"
PASS successfullyParsed is true

TEST COMPLETE
